Use DVFS locks instead of cpufreq policy mangling in pvr.
The pvr driver requires that the memory bus speed is kept fast while the
GPU is active. Since the memory bus speed is tied to the CPU frequency,
pvr (previously) used a cpufreq notifier called on policy updates to ensure
the CPU is running at a minimum 200 MHz while active. The problem with
this approach is that any changes to scaling_min_freq while the GPU is
active (e.g., while interacting with the Processor settings menu) get
clamped to 200 MHz, which persists even when the GPU is inactive.

By switching to DVFS locks, changes to scaling_min_freq persist, and so
allows the use of 100 MHz when the GPU is inactive. Fixes epiccm #14.

Change-Id: I87da92c3b69ae62e78b888080acad9f1c0d99f61